requirements.txt
autopep8==1.6.0
click==8.0.3
pycodestyle==2.8.0
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
pyserial==3.5
PySide6==6.2.2.1
python-dotenv==0.19.2
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
shiboken6==6.2.2.1
toml==0.10.2
